remove compiler warning

+++ utilities/example-2d.c
@@ -102,8 +102,8 @@ int main (int argc, char* argv[])
     /* Find the maximum difference */
     double max = 0.0;
     for (size_t i = 0; i < npoints; i++)
-      if (fabsf(comp[i] - data[i]) > max)
-        max = fabsf(comp[i] - data[i]);
+      if (fabs(comp[i] - data[i]) > max)
+        max = fabs(comp[i] - data[i]);
 
     printf("max diff = %g\n", max);
